# Optimize `ULID.parse` and change exception type for upper `ULID.max`