--- a/sig/ulid.rbs
+++ b/sig/ulid.rbs
@@ -15,6 +15,8 @@ class ULID
   PATTERN: Regexp
   STRICT_PATTERN: Regexp
   UUIDV4_PATTERN: Regexp
+  REPLACING_MAP: Hash[String, String]
+  REPLACING_PATTERN: Regexp
   MONOTONIC_GENERATOR: MonotonicGenerator
   MIN: ULID
   MAX: ULID
@@ -30,6 +32,9 @@ class ULID
   end
 
   class ParserError < Error
+  end
+
+  class SetupError < ScriptError
   end
 
   class MonotonicGenerator
@@ -80,6 +85,7 @@ class ULID
                 | (String string) { (ULID ulid) -> void } -> singleton(ULID)
   def self.octets_from_integer: (Integer integer, length: Integer) -> Array[Integer]
   def self.inverse_of_digits: (Array[Integer] reversed_digits) -> Integer
+  def self.convert_crockford_base32_to_n32: (String) -> String
   attr_reader milliseconds: Integer
   attr_reader entropy: Integer
   def initialize: (milliseconds: Integer, entropy: Integer) -> void