--- a/test/test_many_data.rb
+++ b/test/test_many_data.rb
@@ -17,4 +17,11 @@ class TestManyData < Test::Unit::TestCase
     assert_equal(true, (5..50).cover?(ulids.group_by(&:to_time).size))
     assert_not_equal(ulids, ulids.sort_by(&:to_s))
   end
+
+  def test_parse_reversible
+    10000.times do
+      ulid_string = ULID.from_integer(SecureRandom.random_number(ULID::MAX_INTEGER)).to_s
+      assert_equal(ulid_string, ULID.parse(ulid_string).to_s)
+    end
+  end
 end