For a parser part of #7

```console
$ ruby -v
ruby 3.0.1p64 (2021-04-05 revision 0fb782ee38) [x86_64-darwin20]

$ bundle exec ruby benchmark/parse.rb
Warming up --------------------------------------
ULID.parse_with_integer_base / Before #7
                         1.103k i/100ms
ULID.parse / After #7
                        13.064k i/100ms
Calculating -------------------------------------
ULID.parse_with_integer_base / Before #7
                         12.282k (± 4.2%) i/s -     61.768k in   5.037999s
ULID.parse / After #7
                        129.000k (± 3.1%) i/s -    653.200k in   5.068698s

Comparison:
ULID.parse / After #7:   128999.6 i/s
ULID.parse_with_integer_base / Before #7:    12282.2 i/s - 10.50x  (± 0.00) slower
```

And closes #45